public-opinion-detection/public_opinion_detection/requirements.txt
Django==3.1.7
django_redis==4.12.1
numpy==1.20.2
Pillow==8.4.0
PyMySQL==1.0.2
